pyproject.toml
[tool.poetry]
name = "medvoice-ai"
version = "0.1.0"
description = "Grounded medical voice assistant powered by RAG and semantic search."
authors = ["Jesús Alvarez Heredia"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.128.8"
uvicorn = "^0.40.0"
python-dotenv = "^1.2.1"
sqlalchemy = "^2.0.46"
psycopg2-binary = "^2.9.11"
pydantic-settings = "^2.12.0"
psycopg = {extras = ["binary"], version = "^3.3.2"}
alembic = "^1.18.4"
greenlet = "^3.3.1"
pgvector = "^0.4.2"
asyncpg = "^0.31.0"
openai = "^2.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
